+++ test/btyacc/inherit2.tab.c
@@ -197,7 +197,7 @@ extern int YYPARSE_DECL();
 #define INTEGER 260
 #define NAME 261
 #define YYERRCODE 256
-typedef short YYINT;
+typedef int YYINT;
 static const YYINT inherit2_lhs[] = {                    -1,
     5,    6,    0,    0,    3,    3,    4,    4,    1,    1,
     7,    2,
